Ghost debug (see desc.)
Tried finding the source of extreme slowness when typing. Assumed it was this fork, but turns out to be a markdown plugin that called a function 44000 times in 3.5 seconds. *cough cough Plasticboy cough* -- keeping some of the changes though, just because why not? Might break some backwards compat, but I've never really cared about that