Theoretically fixes #322
Part of it was already fixed, but backwards search didn't respect
multiline.

[]|<bs> now respects multiline, which completely prevents incorrect pair
deletion of non-multiline pairs, including vim comments over two lines.
As in two separate lines of blank comments would be taken down
incorrectly because it would've been registered as a multiline pair.

Basically an oversight on my end when enabling that deletion type. Bad
line removal, again, was already fixed. Not sure what fixed it, but it's
either the comment regex or one of krasjet's changes